pyproject.toml
[project]
name = "album2video"
version = "0.1.0"
description = "Convert travel photo/video albums into 4K MP4 files for Raspberry Pi playback"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0",
    "Pillow>=10.0",
    "pillow-heif>=0.16",
]

[project.scripts]
album2video = "album2video.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/album2video"]
